+++ internal/presentation/rest/handlers/get_orders/schema.go
@@ -19,7 +19,7 @@ func (h) Parse(ctx *gin.Context) (request, error) {
 type orderItem struct {
 	Number     string  `json:"number"`
 	Status     string  `json:"status"`
-	Accrual    float64 `json:"accrual,omitempty"`
+	Accrual    float32 `json:"accrual,omitempty"`
 	UploadedAt string  `json:"uploaded_at"`
 }
