+++ internal/usecases/withdrawals.go
@@ -6,13 +6,13 @@ import (
 	"github.com/xantinium/gophermart/internal/models"
 )
 
-func (cases *UseCases) CreateWithdrawal(ctx context.Context, order string, sum float64, userID int) error {
+func (cases *UseCases) CreateWithdrawal(ctx context.Context, order string, sum float32, userID int) error {
 	balance, err := cases.GetUserBalance(ctx, userID)
 	if err != nil {
 		return err
 	}
 
-	if balance.AvaliableAccrual < float64(sum) {
+	if balance.AvaliableAccrual < sum {
 		return models.ErrInsufficientBalance
 	}
